[GOJHW]-8/at: add tracing in db and cache

+++ internal/storage/facade.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/opentracing/opentracing-go"
 	"gitlab.ozon.dev/akugnerevich/homework.git/internal/cache"
 	"gitlab.ozon.dev/akugnerevich/homework.git/internal/kafka"
 	"gitlab.ozon.dev/akugnerevich/homework.git/internal/models"
@@ -47,21 +48,32 @@ func NewStorageFacade(pool *pgxpool.Pool, producer kafka.Producer) *storageFacad
 }
 
 func (s storageFacade) AcceptOrder(ctx context.Context, or *models.Order) error {
+	parentSpan := opentracing.SpanFromContext(ctx)
+	dbSpan := parentSpan.Tracer().StartSpan("AcceptOrder", opentracing.ChildOf(parentSpan.Context()))
+	defer dbSpan.Finish()
+
+	dbSpan.LogKV("action", "AcceptOrder", "order_id", or.ID)
+
 	if or.KeepUntilDate.Before(time.Now()) {
+		dbSpan.SetTag("error", true)
+		dbSpan.LogKV("AcceptOrder", "error_message", e.ErrDate)
 		return e.ErrDate
 	}
-
-	if _, t := s.cache.Get(strconv.Itoa(int(or.ID))); t {
+	if _, t := s.cache.Get(ctx, strconv.Itoa(int(or.ID))); t {
+		dbSpan.SetTag("error", true)
+		dbSpan.LogKV("AcceptOrder", "error_message", e.ErrIsConsist)
 		return e.ErrIsConsist
 	}
 	or.State = models.AcceptState
 	or.AcceptTime = time.Now().Unix()
 	err := s.PgRepo.AddToStorage(ctx, or)
 	if err != nil {
+		dbSpan.SetTag("error", true)
+		dbSpan.LogKV("AcceptOrder", "error_message", err.Error())
 		return err
 	}
 
-	s.cache.Set(strconv.Itoa(int(or.ID)), or, 5*time.Second)
+	s.cache.Set(ctx, strconv.Itoa(int(or.ID)), or, 5*time.Second)
 
 	err = s.producer.SendMessage(ctx, *or, models.AcceptEvent)
 	if err != nil {
@@ -72,49 +84,86 @@ func (s storageFacade) AcceptOrder(ctx context.Context, or *models.Order) error 
 }
 
 func (s storageFacade) PlaceOrder(ctx context.Context, ids []uint32) error {
+	parentSpan := opentracing.SpanFromContext(ctx)
+	span := parentSpan.Tracer().StartSpan("PlaceOrder", opentracing.ChildOf(parentSpan.Context()))
+	defer span.Finish()
+
+	span.LogKV("action", "PlaceOrder", "order_ids", ids)
+
 	return s.txManager.RunReadCommited(ctx, func(ctxT context.Context) error {
 		orders, exists := s.PgRepo.GetItems(ctxT, ids)
 		if !exists {
+			span.SetTag("error", true)
+			span.LogKV("PlaceOrder", "error_message", e.ErrNoConsist)
 			return e.ErrNoConsist
 		}
+
 		baseUserID := orders[0].UserID
 		for _, order := range orders {
 			if order.UserID != baseUserID {
+				span.SetTag("error", true)
+				span.LogKV("PlaceOrder", "error_message", e.ErrNotAllIDs)
 				return e.ErrNotAllIDs
 			}
-
-			switch order.State {
-			case models.PlaceState:
-				return fmt.Errorf("Order by id: %d is already placed", order.ID)
-			case models.SoftDelete:
-				return fmt.Errorf("Order by id: %d was deleted", order.ID)
+			if order.State == models.PlaceState {
+				err := fmt.Errorf("Order by id: %d is already placed", order.ID)
+				span.SetTag("error", true)
+				span.LogKV("PlaceOrder", "error_message", err.Error())
+				return err
 			}
-
+			if order.State == models.SoftDelete {
+				err := fmt.Errorf("Order by id: %d was deleted", order.ID)
+				span.SetTag("error", true)
+				span.LogKV("PlaceOrder", "error_message", err.Error())
+				return err
+			}
 			if !order.KeepUntilDate.After(time.Now()) {
-				return fmt.Errorf("Order by id: %d cannot be issued to the customer because the date is invalid", order.ID)
+				err := fmt.Errorf("Order by id: %d cannot be issued to the customer because the date is invalid", order.ID)
+				span.SetTag("error", true)
+				span.LogKV("PlaceOrder", "error_message", err.Error())
+				return err
 			}
 		}
 
 		if err := s.PgRepo.UpdateBeforePlace(ctxT, ids, time.Now()); err != nil {
+			span.SetTag("error", true)
+			span.LogKV("PlaceOrder", "error_message", err.Error())
 			return err
 		}
 
 		err := s.producer.SendMessages(ctx, orders, models.PlaceEvent)
 		if err != nil {
+			span.SetTag("error", true)
+			span.LogKV("PlaceOrder", "error_message", ErrLogEvent.Error())
 			return ErrLogEvent
 		}
 		return nil
 	})
 }
 
-func (s storageFacade) ReturnOrder(ctx context.Context, order *models.Order) error {
+func (s storageFacade) ReturnOrder(ctx context.Context, id uint) error {
+	parentSpan := opentracing.SpanFromContext(ctx)
+	span := parentSpan.Tracer().StartSpan("ReturnOrder", opentracing.ChildOf(parentSpan.Context()))
+	defer span.Finish()
+	span.LogKV("action", "ReturnOrder", "order_id", id)
+
 	return s.txManager.RunReadCommited(ctx, func(ctxTx context.Context) error {
+		order, exists := s.PgRepo.GetItem(ctx, id)
+		if !exists {
+			span.SetTag("error", true)
+			span.LogKV("ReturnOrder", "error_message", e.ErrNoConsist)
+			return e.ErrNoConsist
+		}
 		err := order.CanBeReturned()
 		if err != nil {
+			span.SetTag("error", true)
+			span.LogKV("ReturnOrder", "error_message", err.Error())
 			return err
 		}
 		err = s.PgRepo.DeleteFromStorage(ctxTx, order.ID)
 		if err != nil {
+			span.SetTag("error", true)
+			span.LogKV("ReturnOrder", "error_message", err.Error())
 			return err
 		}
 
@@ -122,6 +171,8 @@ func (s storageFacade) ReturnOrder(ctx context.Context, order *models.Order) err
 
 		err = s.producer.SendMessage(ctx, *order, models.ReturnEvent)
 		if err != nil {
+			span.SetTag("error", true)
+			span.LogKV("ReturnOrder", "error_message", ErrLogEvent.Error())
 			return ErrLogEvent
 		}
 
@@ -129,25 +180,47 @@ func (s storageFacade) ReturnOrder(ctx context.Context, order *models.Order) err
 	})
 }
 
-func (s storageFacade) RefundOrder(ctx context.Context, order *models.Order, userId uint) error {
+func (s storageFacade) RefundOrder(ctx context.Context, id uint, userId uint) error {
+	parentSpan := opentracing.SpanFromContext(ctx)
+	span := parentSpan.Tracer().StartSpan("RefundOrder", opentracing.ChildOf(parentSpan.Context()))
+	defer span.Finish()
+
+	span.LogKV("action", "RefundOrder", "order_id", id, "user_id", userId)
+
 	return s.txManager.RunReadCommited(ctx, func(ctxTx context.Context) error {
+		order, exists := s.PgRepo.GetItem(ctx, id)
+		if !exists {
+			span.SetTag("error", true)
+			span.LogKV("RefundOrder", "error_message", e.ErrCheckOrderID)
+			return e.ErrCheckOrderID
+		}
 		if order.State != models.PlaceState {
+			span.SetTag("error", true)
+			span.LogKV("RefundOrder", "error_message", e.ErrNotPlace)
 			return e.ErrNotPlace
 		}
 		if time.Now().After(order.PlaceDate.AddDate(0, 0, 2)) {
+			span.SetTag("error", true)
+			span.LogKV("RefundOrder", "error_message", e.ErrTimeExpired)
 			return e.ErrTimeExpired
 		}
 		if order.UserID != userId {
+			span.SetTag("error", true)
+			span.LogKV("RefundOrder", "error_message", e.ErrIncorrectUserId)
 			return e.ErrIncorrectUserId
 		}
 
 		err := s.PgRepo.UpdateState(ctxTx, order.ID, models.RefundedState)
 		if err != nil {
+			span.SetTag("error", true)
+			span.LogKV("RefundOrder", "error_message", err.Error())
 			return err
 		}
 
 		err = s.producer.SendMessage(ctx, *order, models.RefundEvent)
 		if err != nil {
+			span.SetTag("error", true)
+			span.LogKV("RefundOrder", "error_message", ErrLogEvent.Error())
 			return ErrLogEvent
 		}
 
@@ -156,16 +229,24 @@ func (s storageFacade) RefundOrder(ctx context.Context, order *models.Order, use
 }
 
 func (s storageFacade) ListOrders(ctx context.Context, id uint, inPuP bool, count int) ([]models.Order, error) {
-	cacheKey := fmt.Sprintf("orders_%d_%t_%d", id, inPuP, count)
+	parentSpan := opentracing.SpanFromContext(ctx)
+	span := parentSpan.Tracer().StartSpan("ListOrders", opentracing.ChildOf(parentSpan.Context()))
+	defer span.Finish()
 
-	if cachedOrders, found := s.cache.Get(cacheKey); found {
+	span.LogKV("action", "ListOrders", "user_id", id, "in_progress", inPuP, "count", count)
+
+	cacheKey := fmt.Sprintf("orders_%d_%t_%d", id, inPuP, count)
+	if cachedOrders, found := s.cache.Get(ctx, cacheKey); found {
 		return cachedOrders.([]models.Order), nil
 	}
 
 	list, err := s.PgRepo.GetUserOrders(ctx, id, inPuP)
 	if err != nil {
+		span.SetTag("error", true)
+		span.LogKV("ListOrders", "error_message", err.Error())
 		return nil, err
 	}
+
 	SortOrders(list)
 
 	if count < 1 {
@@ -178,24 +259,31 @@ func (s storageFacade) ListOrders(ctx context.Context, id uint, inPuP bool, coun
 		list = list[:count]
 	}
 
-	s.cache.Set(cacheKey, list, 5*time.Second)
+	s.cache.Set(ctx, cacheKey, list, 5*time.Second)
 
 	return list, nil
 }
 
 func (s storageFacade) ListReturns(ctx context.Context, limit, page int) ([]models.Order, error) {
-	cacheKey := fmt.Sprintf("returns_page_%d_limit_%d", page, limit)
+	parentSpan := opentracing.SpanFromContext(ctx)
+	span := parentSpan.Tracer().StartSpan("ListReturns", opentracing.ChildOf(parentSpan.Context()))
+	defer span.Finish()
 
-	if cachedList, found := s.cache.Get(cacheKey); found {
+	span.LogKV("action", "ListReturns", "page", page, "limit", limit)
+
+	cacheKey := fmt.Sprintf("returns_page_%d_limit_%d", page, limit)
+	if cachedList, found := s.cache.Get(ctx, cacheKey); found {
 		return cachedList.([]models.Order), nil
 	}
 
 	list, err := s.PgRepo.GetReturns(ctx, page-1, limit)
 	if err != nil {
+		span.SetTag("error", true)
+		span.LogKV("ListReturns", "error_message", err.Error())
 		return nil, err
 	}
 
-	s.cache.Set(cacheKey, list, 5*time.Second)
+	s.cache.Set(ctx, cacheKey, list, 5*time.Second)
 
 	return list, nil
 }
